Report real compaction counts instead of a hardcoded zero
`/status` always printed `compactions: 0` because `Transcript::compaction_count`
returned a literal `0`. The same stub fed the status bar's compaction segment,
which is gated on `> 0`, so that indicator could never appear at all. A
benchmark measured the stub, so it timed a function returning a constant.

A second defect had the same root: the backend projects a compaction
checkpoint as a `system` role message, and the UI's hydration match dropped
that role silently, so a resumed session showed no sign compaction had
happened even though the summary was in the log.

The count is now backend-owned. `SessionStats` carries an optional
`compaction_count` derived from `CompactionCheckpoint` events in the session
log, which keeps it correct live and after resume, and unaffected by a
client-side `/clear` that must not erase a session fact. This needed no new
event class, only an additive optional field.

The UI gains a `Compaction` transcript entry kind so checkpoints render with
their summary rather than vanishing.

Token totals are deliberately not included. `SessionStats.total_tokens` is
still `None`: the runner's usage accumulator is per-turn by construction --
its own comment notes each request bills its own input -- and no session
event persists provider usage, so there is nothing to sum across a session
without new durable state. Displaying the per-turn figure as a session total
would be wrong. Recorded as a follow-up rather than guessed at.

Tests: the projection reports 0 then 2 as checkpoints accumulate, verified to
fail before this change; a compaction entry renders its summary.

+++ crates/yach-ui/src/layout.rs
@@ -20,7 +20,7 @@ pub struct RenderParams<'a> {
     pub approval_mode: &'a str,
     pub status_message: &'a str,
     pub is_connected: bool,
-    pub compaction_count: usize,
+    pub compaction_count: u64,
     pub context_used_percent: Option<u8>,
     pub context_window: Option<u64>,
     pub terminal_focused: bool,
